+++ tests/value_test.cc
@@ -4,8 +4,6 @@
 #include "webcface/field.h"
 #include <webcface/member.h>
 #include <webcface/value.h>
-#include <webcface/text.h>
-#include <webcface/log.h>
 #include <stdexcept>
 #include <chrono>
 
@@ -14,7 +12,7 @@ static SharedString operator""_ss(const char *str, std::size_t len) {
     return SharedString::fromU8String(std::string_view(str, len));
 }
 
-class DataTest : public ::testing::Test {
+class ValueTest : public ::testing::Test {
   protected:
     void SetUp() override {
         data_ = std::make_shared<internal::ClientData>(self_name);
@@ -41,21 +39,6 @@ class DataTest : public ::testing::Test {
     Value value(const T1 &member, const T2 &name) {
         return Value{field(member, name)};
     }
-    template <typename T1, typename T2>
-    Text text(const T1 &member, const T2 &name) {
-        return Text{field(member, name)};
-    }
-    template <typename T1, typename T2>
-    Variant variant(const T1 &member, const T2 &name) {
-        return Variant{field(member, name)};
-    }
-    Log log(const SharedString &member, const SharedString &name) {
-        return Log{Field{data_, member, name}};
-    }
-    Log log(std::string_view member, std::string_view name) {
-        return Log{Field{data_, SharedString::fromU8String(member),
-                         SharedString::fromU8String(name)}};
-    }
     int callback_called;
     template <typename V>
     auto callback() {
@@ -66,7 +49,7 @@ class DataTest : public ::testing::Test {
     }
 };
 
-TEST_F(DataTest, field) {
+TEST_F(ValueTest, field) {
     EXPECT_EQ(value("a", "b").member().name(), "a");
     EXPECT_EQ(value("a", "b").member().nameW(), L"a");
     EXPECT_EQ(value("a", "b").name(), "b");
@@ -74,18 +57,10 @@ TEST_F(DataTest, field) {
     EXPECT_EQ(value("a", "b").child("c").name(), "b.c");
     EXPECT_EQ(value("a", "b").child(L"c").name(), "b.c");
     EXPECT_EQ(value("a", "b.c").parent().name(), "b");
-    EXPECT_EQ(text("a", "b").member().name(), "a");
-    EXPECT_EQ(text("a", "b").name(), "b");
-    EXPECT_EQ(text("a", "b").nameW(), L"b");
-    EXPECT_EQ(text("a", "b").child("c").name(), "b.c");
-    EXPECT_EQ(text("a", "b").child(L"c").name(), "b.c");
-    EXPECT_EQ(log("a", "b").member().name(), "a");
 
     EXPECT_THROW(Value().tryGet(), std::runtime_error);
-    EXPECT_THROW(Text().tryGet(), std::runtime_error);
-    EXPECT_THROW(Log().tryGet(), std::runtime_error);
 }
-TEST_F(DataTest, eventTarget) {
+TEST_F(ValueTest, eventTarget) {
     value("a", "b").onChange(callback<Value>());
     data_->value_change_event["a"_ss]["b"_ss]->operator()(field("a", "b"));
     EXPECT_EQ(callback_called, 1);
@@ -96,18 +71,8 @@ TEST_F(DataTest, eventTarget) {
     data_->value_change_event["a"_ss]["b"_ss]->operator()(field("a", "b"));
     EXPECT_EQ(callback_called, 1);
     callback_called = 0;
-
-    text("a", "b").onChange(callback<Text>());
-    data_->text_change_event["a"_ss]["b"_ss]->operator()(field("a", "b"));
-    EXPECT_EQ(callback_called, 1);
-    callback_called = 0;
-
-    log("a", "b").onChange(callback<Log>());
-    data_->log_append_event["a"_ss]["b"_ss]->operator()(field("a"));
-    EXPECT_EQ(callback_called, 1);
-    callback_called = 0;
 }
-TEST_F(DataTest, valueSet) {
+TEST_F(ValueTest, valueSet) {
     data_->value_change_event[self_name]["b"_ss] =
         std::make_shared<std::function<void(Value)>>(callback<Value>());
     value(self_name, "b").set(123);
@@ -115,7 +80,7 @@ TEST_F(DataTest, valueSet) {
     EXPECT_EQ(callback_called, 1);
     EXPECT_THROW(value("a", "b").set(123), std::invalid_argument);
 }
-TEST_F(DataTest, valueSetVec) {
+TEST_F(ValueTest, valueSetVec) {
     data_->value_change_event[self_name]["d"_ss] =
         std::make_shared<std::function<void(Value)>>(callback<Value>());
     value(self_name, "d").set({1, 2, 3, 4, 5});
@@ -152,7 +117,7 @@ TEST_F(DataTest, valueSetVec) {
     EXPECT_EQ((*data_->value_store.getRecv(self_name, "d7"_ss)).size(), 5u);
     EXPECT_EQ((*data_->value_store.getRecv(self_name, "d8"_ss)).size(), 5u);
 }
-// TEST_F(DataTest, ArrayLike){
+// TEST_F(ValueTest, ArrayLike){
 static_assert(
     std::is_same_v<traits::ElementTypeOf<std::vector<double>>, const double &>);
 static_assert(std::is_same_v<traits::ElementTypeOf<std::array<double, 5>>,
@@ -252,39 +217,7 @@ static_assert(!traits::NestedArraySizeMatch<
 static_assert(!traits::NestedArraySizeMatch<double (&)[5][5], 5>::value);
 
 
-TEST_F(DataTest, textSet) {
-    data_->text_change_event[self_name]["b"_ss] =
-        std::make_shared<std::function<void(Variant)>>(callback<Variant>());
-    text(self_name, "b").set("c");
-    EXPECT_EQ(
-        static_cast<std::string>(*data_->text_store.getRecv(self_name, "b"_ss)),
-        "c");
-    EXPECT_EQ(callback_called, 1);
-    EXPECT_THROW(text("a", "b").set("c"), std::invalid_argument);
-}
-TEST_F(DataTest, textSetW) {
-    data_->text_change_event[self_name]["b"_ss] =
-        std::make_shared<std::function<void(Variant)>>(callback<Variant>());
-    text(self_name, "b").set(L"c");
-    EXPECT_EQ(
-        static_cast<std::string>(*data_->text_store.getRecv(self_name, "b"_ss)),
-        "c");
-    EXPECT_EQ(callback_called, 1);
-    EXPECT_THROW(text("a", "b").set(L"c"), std::invalid_argument);
-}
-// TEST_F(DataTest, variantSet) {
-//     data_->text_change_event[self_name]["b"_ss] =
-//         std::make_shared<std::function<void(Variant)>>(callback<Variant>());
-//     variant(self_name, "b").set(ValAdaptor(123));
-//     EXPECT_EQ(static_cast<std::string>(
-//                   *data_->text_store.getRecv(self_name, "b"_ss)),
-//               "123");
-//     EXPECT_EQ(callback_called, 1);
-//     EXPECT_THROW(variant("a", "b").set(ValAdaptor(123)),
-//     std::invalid_argument);
-// }
-
-TEST_F(DataTest, valueGet) {
+TEST_F(ValueTest, valueGet) {
     data_->value_store.setRecv(
         "a"_ss, "b"_ss,
         std::make_shared<std::vector<double>>(std::vector<double>({123})));
@@ -299,73 +232,3 @@ TEST_F(DataTest, valueGet) {
     value("a", "d").onChange(callback<Value>());
     EXPECT_EQ(data_->value_store.transferReq().at("a"_ss).at("d"_ss), 3u);
 }
-TEST_F(DataTest, textGet) {
-    data_->text_store.setRecv("a"_ss, "b"_ss,
-                              std::make_shared<ValAdaptor>("hoge"));
-    ASSERT_NE(text("a", "b").tryGet(), std::nullopt);
-    EXPECT_EQ(text("a", "b").tryGet().value(), "hoge");
-    EXPECT_EQ(text("a", "b").tryGetW().value(), L"hoge");
-    EXPECT_EQ(variant("a", "b").tryGet().value(), ValAdaptor("hoge"));
-    EXPECT_EQ(text("a", "b").get(), "hoge");
-    EXPECT_EQ(text("a", "b").getW(), L"hoge");
-    EXPECT_EQ(variant("a", "b").get().asStringRef(), "hoge");
-    EXPECT_EQ(text("a", "c").tryGet(), std::nullopt);
-    EXPECT_EQ(text("a", "c").tryGetW(), std::nullopt);
-    EXPECT_EQ(variant("a", "c").tryGet(), std::nullopt);
-    EXPECT_EQ(text("a", "c").get(), "");
-    EXPECT_EQ(text("a", "c").getW(), L"");
-    EXPECT_TRUE(variant("a", "c").get().empty());
-    EXPECT_EQ(data_->text_store.transferReq().at("a"_ss).at("b"_ss), 1u);
-    EXPECT_EQ(data_->text_store.transferReq().at("a"_ss).at("c"_ss), 2u);
-    EXPECT_EQ(text(self_name, "b").tryGet(), std::nullopt);
-    EXPECT_EQ(data_->text_store.transferReq().count(self_name), 0u);
-    text("a", "d").onChange(callback<Text>());
-    EXPECT_EQ(data_->text_store.transferReq().at("a"_ss).at("d"_ss), 3u);
-}
-TEST_F(DataTest, logGet) {
-    using namespace std::chrono;
-    auto logs = std::make_shared<webcface::internal::LogHistory>(
-        std::deque<LogLineData>{
-            {1, system_clock::now(), "a"_ss},
-            {2, system_clock::now(), "b"_ss},
-            {3, system_clock::now(), "c"_ss},
-        });
-    data_->log_store.setRecv("a"_ss, "b"_ss, logs);
-    EXPECT_EQ(log("a", "b").tryGet().value().size(), 3u);
-    EXPECT_EQ(log("a", "b").tryGetW().value().size(), 3u);
-    ASSERT_EQ(log("a", "b").get().size(), 3u);
-    ASSERT_EQ(log("a", "b").getW().size(), 3u);
-    EXPECT_EQ(log("a", "b").get()[2].level(), 3);
-    EXPECT_EQ(log("a", "b").getW()[2].level(), 3);
-    EXPECT_EQ(log("a", "b").get()[2].message(), "c");
-    EXPECT_EQ(log("a", "b").getW()[2].message(), L"c");
-    EXPECT_EQ(log("a", "c").tryGet(), std::nullopt);
-    EXPECT_EQ(log("a", "c").tryGetW(), std::nullopt);
-    EXPECT_EQ(log("a", "c").get().size(), 0u);
-    EXPECT_EQ(log("a", "c").getW().size(), 0u);
-    EXPECT_EQ(data_->log_store.transferReq().at("a"_ss).at("b"_ss), 1u);
-    EXPECT_EQ(data_->log_store.transferReq().at("a"_ss).at("c"_ss), 2u);
-    EXPECT_FALSE(log(self_name, "default"_ss).tryGet().has_value());
-    EXPECT_FALSE(log(self_name, "default"_ss).tryGetW().has_value());
-    EXPECT_FALSE(log(self_name, "a"_ss).tryGet().has_value());
-    EXPECT_FALSE(log(self_name, "a"_ss).tryGetW().has_value());
-    EXPECT_EQ(log(self_name, "a"_ss).get().size(), 0u);
-    EXPECT_EQ(log(self_name, "a"_ss).getW().size(), 0u);
-    EXPECT_EQ(data_->log_store.transferReq().count(self_name), 0u);
-    log("a", "d").onChange(callback<Log>());
-    EXPECT_EQ(data_->log_store.transferReq().at("a"_ss).at("d"_ss), 3u);
-}
-TEST_F(DataTest, logClear) {
-    using namespace std::chrono;
-    auto logs = std::make_shared<webcface::internal::LogHistory>(
-        std::deque<LogLineData>{
-            {1, system_clock::now(), "a"_ss},
-            {2, system_clock::now(), "b"_ss},
-            {3, system_clock::now(), "c"_ss},
-        });
-    data_->log_store.setRecv("a"_ss, "b"_ss, logs);
-    log("a", "b").clear();
-    EXPECT_EQ(log("a", "b").tryGet().value().size(), 0u);
-    EXPECT_EQ(log("a", "b").tryGetW().value().size(), 0u);
-}
-// todo: hidden, free
